Handle unix-socket communication with docker-server
In the MacOS version the cli relied on DOCKER_HOST to be set to a specific IP address (\d+.\d+.\d+.\d+).
When running on a LINUX machine with a local docker daemon, DOCKER_HOST is set to `unix:///var/run/docker.sock`.

If DOCKER_HOST is set like that the CLI extracts the IP address of the host and using ssh with this address.